Add extra check report for bad records
